+++ src/data/threads.js
@@ -15,7 +15,7 @@ const updates = require("./updates");
 const Thread = require("./Thread");
 const ThreadMessage = require("./ThreadMessage");
 const {callBeforeNewThreadHooks} = require("../hooks/beforeNewThread");
-const {THREAD_STATUS, DISOCRD_CHANNEL_TYPES} = require("./constants");
+const {THREAD_STATUS, DISCORD_CHANNEL_TYPES} = require("./constants");
 const {findNotesByUserId} = require("./notes");
 
 const MINUTES = 60 * 1000;
@@ -158,13 +158,12 @@ async function createNewThreadForUser(user, opts = {}) {
       }
     }
 
-    // Use the user's name+discrim for the thread channel's name
+    // Use the user's name for the thread channel's name
     // Channel names are particularly picky about what characters they allow, so we gotta do some clean-up
     let cleanName = transliterate.slugify(user.username);
     if (cleanName === "") cleanName = "unknown";
-    cleanName = cleanName.slice(0, 95); // Make sure the discrim fits
 
-    let channelName = `${cleanName}-${user.discriminator}`;
+    let channelName = cleanName;
 
     if (config.anonymizeChannelName) {
       channelName = crypto.createHash("md5").update(channelName + Date.now()).digest("hex").slice(0, 12);
@@ -206,7 +205,7 @@ async function createNewThreadForUser(user, opts = {}) {
     // Attempt to create the inbox channel for this thread
     let createdChannel;
     try {
-      createdChannel = await utils.getInboxGuild().createChannel(opts.channelName, DISOCRD_CHANNEL_TYPES.GUILD_TEXT, {
+      createdChannel = await utils.getInboxGuild().createChannel(opts.channelName, DISCORD_CHANNEL_TYPES.GUILD_TEXT, {
         reason: "New Modmail thread",
         parentID: newThreadCategoryId,
       });
@@ -215,7 +214,7 @@ async function createNewThreadForUser(user, opts = {}) {
       if (err.message.includes("Contains words not allowed for servers in Server Discovery")) {
         const replacedChannelName = "badname-0000";
         try {
-          createdChannel = await utils.getInboxGuild().createChannel(replacedChannelName, DISOCRD_CHANNEL_TYPES.GUILD_TEXT, {
+          createdChannel = await utils.getInboxGuild().createChannel(replacedChannelName, DISCORD_CHANNEL_TYPES.GUILD_TEXT, {
             reason: "New Modmail thread",
             parentID: newThreadCategoryId,
           });
@@ -233,7 +232,7 @@ async function createNewThreadForUser(user, opts = {}) {
     const newThreadId = await createThreadInDB({
       status: THREAD_STATUS.OPEN,
       user_id: user.id,
-      user_name: `${user.username}#${user.discriminator}`,
+      user_name: user.username,
       channel_id: createdChannel.id,
       next_message_number: 1,
       created_at: moment.utc().format("YYYY-MM-DD HH:mm:ss")